refact: put fileset.nix into the repository. In order to pass currentSystem bug, I thought if I am not changing api often, I can copy the freezeFileset function into snack repository. However, bin/packages.nix also needs currentSystem. Unless I make snack-lib.executable like callPackages { inherit pkgs; }, I can't give bin/pacakges.nix the right pkgs. But snack-lib.executable is supposed to give snack-compiled package a free choice for pkgs.
After I use `nix develop --impure`, the error is resolved.
So.. I probably have done a right thing (make a lib function inside repository) for a wrong reason (want to purify the supposed-to-be-impure).
Anyway, now I can nix develop (flake-based) and get snack in the shell.
Probably the evaluation cache will be more prominent when I snack-compile a larger project.